requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.12.0
    # via
    #   tensorboard
    #   tensorflow
argon2-cffi==20.1.0
    # via notebook
astunparse==1.6.3
    # via tensorflow
async-generator==1.10
    # via nbclient
attrs==20.3.0
    # via jsonschema
backcall==0.2.0
    # via ipython
bleach==3.3.0
    # via nbconvert
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via argon2-cffi
chardet==4.0.0
    # via requests
cycler==0.10.0
    # via matplotlib
decorator==5.0.7
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.3
    # via nbconvert
flatbuffers==1.12
    # via tensorflow
gast==0.3.3
    # via tensorflow
google-auth-oauthlib==0.4.4
    # via tensorboard
google-auth==1.30.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.32.0
    # via
    #   tensorboard
    #   tensorflow
h5py==2.10.0
    # via tensorflow
idna==2.10
    # via requests
ipykernel==5.5.3
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   notebook
    #   qtconsole
    #   traitlets
ipython==7.23.0
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==7.6.3
    # via jupyter
jedi==0.18.0
    # via ipython
jinja2==2.11.3
    # via
    #   nbconvert
    #   notebook
jsonschema==3.2.0
    # via nbformat
jupyter-client==6.1.12
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.0
    # via jupyter
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter==1.0.0
    # via -r requirements.in
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.0
    # via ipywidgets
keras-preprocessing==1.1.2
    # via tensorflow
kiwisolver==1.3.1
    # via matplotlib
markdown==3.3.4
    # via tensorboard
markupsafe==1.1.1
    # via jinja2
matplotlib-inline==0.1.2
    # via ipython
matplotlib==3.4.1
    # via -r requirements.in
mistune==0.8.4
    # via nbconvert
nbclient==0.5.3
    # via nbconvert
nbconvert==6.0.7
    # via
    #   jupyter
    #   notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via nbclient
notebook==6.3.0
    # via
    #   jupyter
    #   widgetsnbextension
numpy==1.19.5
    # via
    #   -r requirements.in
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   opt-einsum
    #   tensorboard
    #   tensorflow
oauthlib==3.1.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==20.9
    # via bleach
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.2.0
    # via matplotlib
prometheus-client==0.10.1
    # via notebook
prompt-toolkit==3.0.18
    # via
    #   ipython
    #   jupyter-console
protobuf==3.15.8
    # via
    #   tensorboard
    #   tensorflow
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via cffi
pygments==2.8.1
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   jupyter-client
    #   matplotlib
pyzmq==22.0.3
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.0.3
    # via jupyter
qtpy==1.9.0
    # via qtconsole
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   requests-oauthlib
    #   tensorboard
rsa==4.7.2
    # via google-auth
send2trash==1.5.0
    # via notebook
six==1.15.0
    # via
    #   absl-py
    #   argon2-cffi
    #   astunparse
    #   bleach
    #   cycler
    #   google-auth
    #   google-pasta
    #   grpcio
    #   h5py
    #   jsonschema
    #   keras-preprocessing
    #   protobuf
    #   python-dateutil
    #   tensorflow
tensorboard-data-server==0.6.0
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard==2.5.0
    # via tensorflow
tensorflow-estimator==2.4.0
    # via tensorflow
tensorflow==2.4.1
    # via -r requirements.in
termcolor==1.1.0
    # via tensorflow
terminado==0.9.4
    # via notebook
testpath==0.4.4
    # via nbconvert
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.60.0
    # via -r requirements.in
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==3.7.4.3
    # via tensorflow
urllib3==1.26.4
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
werkzeug==1.0.1
    # via tensorboard
wheel==0.36.2
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
widgetsnbextension==3.5.1
    # via ipywidgets
wrapt==1.12.1
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
